src/pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "aibulat.ryic", email = "aibulat.ryic@gmail.com"},
]
dependencies = [
    "fastapi~=0.70",
    "uvicorn~=0.15",
    "mypy~=0.910",
    "google-cloud-texttospeech~=2.7",
    "aiofiles~=0.8",
    "pydantic~=1.8",
    "python-multipart~=0.0",
    "google-cloud-translate~=3.6",
    "uvloop~=0.16",
]
requires-python = ">=3.9"
dynamic = ["classifiers"]
license = {text = "MIT"}

[project.urls]
homepage = ""

[build-system]
requires = ["pdm-pep517"]
build-backend = "pdm.pep517.api"

[tool]
[tool.pdm]
